backend/chat/serializers.py
from rest_framework import serializers
from .models import Conversation, Message


class ConversationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Conversation
        fields = [
            'id',
            'supplier',
            'consumer',
            'order',
            'conversation_type',
            'created_by',
            'created_at',
            'updated_at',
        ]
        read_only_fields = ['created_by', 'created_at', 'updated_at']


class MessageSerializer(serializers.ModelSerializer):
    sender_username = serializers.CharField(source='sender.username', read_only=True)

    class Meta:
        model = Message
        fields = [
            'id',
            'conversation',
            'sender',
            'sender_username',
            'text',
            'attachment',
            'sent_at',
            'is_read',
        ]
        read_only_fields = ['sender', 'sent_at', 'is_read', 'sender_username']
